replaced main with newmain
